requirements.txt
fastapi
uvicorn
spacy
nltk
sentence-transformers
faiss-cpu
pydantic
python-multipart